+++ demo/index.js
@@ -78,16 +78,35 @@ window.vditor = new Vditor('vditor', {
       'sd': '💔',
       'j': 'https://unpkg.com/vditor@1.3.1/dist/images/emoji/j.png',
     },
-    at: (key) => {
-      if ('vanessa'.indexOf(key.toLocaleLowerCase()) > -1) {
-        return [
-          {
-            value: '@Vanessa',
-            html: '<img src="https://avatars0.githubusercontent.com/u/970828?s=60&v=4"/> Vanessa',
-          }]
-      }
-      return []
-    },
+    extend: [
+      {
+        key: '@',
+        hint: (key) => {
+          console.log(key)
+          if ('vanessa'.indexOf(key.toLocaleLowerCase()) > -1) {
+            return [
+              {
+                value: '@Vanessa',
+                html: '<img src="https://avatars0.githubusercontent.com/u/970828?s=60&v=4"/> Vanessa',
+              }]
+          }
+          return []
+        },
+      },
+      {
+        key: '#',
+        hint: (key) => {
+          console.log(key)
+          if ('vditor'.indexOf(key.toLocaleLowerCase()) > -1) {
+            return [
+              {
+                value: '#Vditor',
+                html: '<span style="color: #999;">#Vditor</span> ♏ 一款浏览器端的 Markdown 编辑器，支持所见即所得（富文本）、即时渲染（类似 Typora）和分屏预览模式。',
+              }]
+          }
+          return []
+        },
+      }],
   },
   tab: '\t',
   upload: {