+++ src/ts/util/editorCommonEvent.ts
@@ -85,7 +85,8 @@ export const scrollCenter = (vditor: IVditor) => {
 export const hotkeyEvent = (vditor: IVditor, editorElement: HTMLElement) => {
     editorElement.addEventListener("keydown", (event: KeyboardEvent & { target: HTMLElement }) => {
         // hint: 上下选择
-        if ((vditor.options.hint.at || vditor.toolbar.elements.emoji) && vditor.hint.select(event, vditor)) {
+        if ((vditor.options.hint.extend.length > 1 || vditor.toolbar.elements.emoji) &&
+            vditor.hint.select(event, vditor)) {
             return;
         }