+++ CHANGELOG.md
@@ -79,7 +79,7 @@
 
 ### v3.4.2 / 2020-08-xx
 
-* [411](https://github.com/Vanessa219/vditor/issues/411) 对标 Typora 反馈 `改进功能`
+* [441](https://github.com/Vanessa219/vditor/issues/441) 对标 Typora 反馈 `改进功能`
 * [669](https://github.com/Vanessa219/vditor/issues/669) links problem(Wysiwyg) `改进功能`
 * [667](https://github.com/Vanessa219/vditor/issues/667) lists (SV) `修复缺陷`
 * [446](https://github.com/Vanessa219/vditor/issues/446) 改进复制粘贴链接体验 `引入特性`